Switched libraries definition into hook implementation.
This change aims to make the Nexx integration compatible with lazy loading of content.
The library definition has now been moved into the hook implementation.
This enables the build of the external Nexx resource, which itself requires the Omnia ID as an argument.

Because the config object isn't needed inside the field formatter anymore, it has been completely removed from the corresponding class.